Conflicts:
	Projects/ClientApp/Src/ClientApplication.cpp
	Projects/Gameplay/Src/Gameplay.vcxproj.filters
	Projects/Gameplay/Src/Transform.cpp
	Projects/Gameplay/Src/Transform.h
	Projects/Graphics/Src/Graphics.vcxproj.filters

+++ Projects/Gameplay/Src/GraphicsBackendSystem.cpp
@@ -167,5 +167,5 @@ void GraphicsBackendSystem::renderAParticleSystem(AglParticleSystem* p_system){
 }
 
 AglVector2 GraphicsBackendSystem::getWindowSize(){
-	return AglVector2(m_scrWidth, m_scrHeight);
+	return AglVector2((float)m_scrWidth, (float)m_scrHeight);
 }
